Added JSONata examples
