--- src/main/java/com/api/jsonata4java/TesterUiPerferences.java
@@ -1,421 +1,0 @@
-package com.api.jsonata4java;
-
-import java.awt.BorderLayout;
-import java.awt.GridBagConstraints;
-import java.awt.GridBagLayout;
-import java.awt.Insets;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
-import java.awt.event.WindowAdapter;
-import java.awt.event.WindowEvent;
-import java.io.File;
-
-import javax.swing.JButton;
-import javax.swing.JDialog;
-import javax.swing.JFileChooser;
-import javax.swing.JFrame;
-import javax.swing.JLabel;
-import javax.swing.JPanel;
-import javax.swing.JTextField;
-import javax.swing.filechooser.FileFilter;
-
-public class TesterUiPerferences extends JDialog {
-
-	private final TesterUI ui;
-
-	private final TesterUISettings settings;
-
-	public TesterUiPerferences(final TesterUI ui, final TesterUISettings settings) {
-		this.ui = ui;
-		this.settings = settings;
-	}
-
-	private static final int FRAME_SIZE_Y = 600;
-
-	private static final int FRAME_SIZE_X = 800;
-
-	private final JFrame frame = new JFrame("JSONata4Java Tester UI Settings");
-
-//	private final JComboBox<String> driverKind = new JComboBox<String>(new String[] {
-//			"A",
-//			"B",
-//	});
-
-	private static final String DOTS_3 = "...";
-
-	private final JTextField pathInput = new JTextField();
-	private final JButton changePathInput = new JButton(DOTS_3);
-	private final JButton loadInput = new JButton("Load");
-	private final JTextField pathJsonata = new JTextField();
-	private final JButton changePathJsonata = new JButton(DOTS_3);
-	private final JButton loadJsonata = new JButton("Load");
-
-//	private final JTextField driverFilePathFi = new JTextField();
-//
-//	private final JButton changeDriverFilePathFi = new JButton(DOTS_3);
-//
-//	private final JTextField driverFilePathCh = new JTextField();
-//
-//	private final JButton changeDriverFilePathCh = new JButton(DOTS_3);
-//
-//	private final JTextField downloadFolder = new JTextField();
-//
-//	private final JButton changeDownloadFolder = new JButton(DOTS_3);
-//
-//	private final JTextField driverMaxWaitSeconds = new JTextField();
-//
-//	private final JTextField driverMaxWaitSecondsPageLoad = new JTextField();
-//
-//	private final JTextField driverMaxWaitSecondsScript = new JTextField();
-//
-//	private final JTextField driverMaxWaitSecondsStart = new JTextField();
-//
-//	private final JTextField driverMaxWaitSecondsStop = new JTextField();
-//
-//	private final JTextField driverMaxWaitSecondsUpload = new JTextField();
-//
-//	private final JCheckBox advancedModeCheckBox = new JCheckBox();
-
-	private final JPanel inputPanel = new JPanel();
-
-	private final JPanel southPanel = new JPanel();
-
-	private final JPanel buttonsPanel = new JPanel();
-
-	private final JButton okButton = new JButton("Ok");
-
-	private final JButton cancelButton = new JButton("Cancel");
-
-	private static final int PAD = 5;
-
-	private static final Insets INSETS = new Insets(PAD, PAD, PAD, PAD);
-
-	private static final long serialVersionUID = 1L;
-
-	public void open() {
-		readProps();
-		initUi();
-		frame.setVisible(true);
-	}
-
-	private void applyAndClose() {
-		writeProps();
-		close();
-	}
-
-	private void changeFilePath(final JTextField textField, final String description, final String... suffix) {
-		final JFileChooser fc = new JFileChooser();
-		final String path = textField.getText();
-		final File fileChosenBefore = path != null && path.trim().length() > 0 ? new File(path) : null;
-		if (fileChosenBefore != null && fileChosenBefore.getParentFile().exists()) {
-			fc.setCurrentDirectory(fileChosenBefore.getParentFile());
-		}
-		fc.setFileFilter(new FileFilter() {
-			@Override
-			public String getDescription() {
-				return description;
-			}
-
-			@Override
-			public boolean accept(final File f) {
-				return acceptFile(f, suffix);
-			}
-		});
-		int ret = fc.showOpenDialog(frame.getContentPane());
-		switch (ret) {
-		case 0:
-			textField.setText(fc.getSelectedFile().getAbsolutePath());
-			break;
-		default:
-			break;
-		}
-	}
-
-	private boolean acceptFile(final File f, final String... suffix) {
-		if (f.isDirectory()) {
-			return true;
-		}
-		try {
-			for (int i = 0; i < suffix.length; i++) {
-				if (f.getName().endsWith(suffix[i])) {
-					return true;
-				}
-			}
-			return false;
-		} catch (StringIndexOutOfBoundsException e) {
-			return false;
-		}
-	}
-
-	private void readProps() {
-		pathInput.setText(settings.getPathInput().toString());
-		pathJsonata.setText(settings.getPathJsonata().toString());
-//		driverKind.setSelectedItem(props.getDriverKind());
-//		driverFilePath.setText(props.getDriverFilePath());
-//		driverFilePathFp.setText(props.getDriverFilePathFirefoxPortable());
-//		driverFilePathFi.setText(props.getDriverFilePathFirefox());
-//		driverFilePathCh.setText(props.getDriverFilePathChrome());
-//		downloadFolder.setText(props.getDownloadFolder());
-//		earUploadMode.setSelectedItem(props.getEarUploadMode());
-//		driverMaxWaitSeconds.setText(props.getDriverMaxWaitSeconds());
-//		driverMaxWaitSecondsPageLoad.setText(props.getDriverMaxWaitSecondsPageLoad());
-//		driverMaxWaitSecondsScript.setText(props.getDriverMaxWaitSecondsScript());
-//		driverMaxWaitSecondsStart.setText(props.getDriverMaxWaitSecondsStart());
-//		driverMaxWaitSecondsStop.setText(props.getDriverMaxWaitSecondsStop());
-//		driverMaxWaitSecondsUpload.setText(props.getDriverMaxWaitSecondsUpload());
-//		advancedModeCheckBox.setSelected(Boolean.parseBoolean(props.getAdvancedMode()));
-	}
-
-	private void writeProps() {
-		settings.setPathInput(new File(pathInput.getText()).toPath());
-		settings.setPathJsonata(new File(pathJsonata.getText()).toPath());
-//		props.setDriverKind(driverKind.getSelectedItem().toString());
-//		props.setDriverFilePath(driverFilePath.getText());
-//		props.setDriverFilePathFirefoxPortable(driverFilePathFp.getText());
-//		props.setDriverFilePathFirefox(driverFilePathFi.getText());
-//		props.setDriverFilePathChrome(driverFilePathCh.getText());
-//		props.setDownloadFolder(downloadFolder.getText());
-//		props.setEarUploadMode(earUploadMode.getSelectedItem().toString());
-//		props.setDriverMaxWaitSeconds(driverMaxWaitSeconds.getText());
-//		props.setDriverMaxWaitSecondsPageLoad(driverMaxWaitSecondsPageLoad.getText());
-//		props.setDriverMaxWaitSecondsScript(driverMaxWaitSecondsScript.getText());
-//		props.setDriverMaxWaitSecondsStart(driverMaxWaitSecondsStart.getText());
-//		props.setDriverMaxWaitSecondsStop(driverMaxWaitSecondsStop.getText());
-//		props.setDriverMaxWaitSecondsUpload(driverMaxWaitSecondsUpload.getText());
-//		props.setAdvancedMode(Boolean.toString(advancedModeCheckBox.isSelected()));
-//		parentApp.updateUiAdvancedMode();
-	}
-
-	private void cancel() {
-		close();
-	}
-
-	private void close() {
-		frame.setVisible(false);
-		frame.dispose();
-	}
-
-	private void initUi() {
-		uiConfigure();
-		uiSetListeners();
-		uiCompose();
-	}
-
-	private void uiConfigure() {
-		frame.setLayout(new BorderLayout(0, 0));
-		frame.setSize(FRAME_SIZE_X, FRAME_SIZE_Y);
-		frame.setLocationRelativeTo(null);
-		frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-		// frame.setIconImage(parentApp.getIconApp().getImage());
-		inputPanel.setLayout(new GridBagLayout());
-		southPanel.setLayout(new BorderLayout());
-		buttonsPanel.setLayout(new GridBagLayout());
-	}
-
-	private void uiCompose() {
-		int x = 1;
-		int y = 1;
-
-//		inputPanel.add(new JLabel("Selenium Driver Kind"), new GridBagConstraints(1, y, 1, 1, 0.0, 0.0,
-//				GridBagConstraints.WEST, GridBagConstraints.NONE, INSETS, 0, 0));
-//		inputPanel.add(driverKind, new GridBagConstraints(2, y++, 1, 1, 1.0, 1.0,
-//				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-
-		x = 1;
-		inputPanel.add(new JLabel("Input JSON / XML Path"), new GridBagConstraints(x++, y, 1, 1, 0.0, 0.0,
-				GridBagConstraints.WEST, GridBagConstraints.NONE, INSETS, 0, 0));
-		inputPanel.add(pathInput, new GridBagConstraints(x++, y, 1, 1, 1.0, 1.0,
-				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-		inputPanel.add(changePathInput, new GridBagConstraints(x++, y, 1, 1, 0.0, 0.0,
-				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-		inputPanel.add(loadInput, new GridBagConstraints(x++, y++, 1, 1, 0.0, 0.0,
-				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-
-		x = 1;
-		inputPanel.add(new JLabel("JSONata Expression Path"), new GridBagConstraints(x++, y, 1, 1, 0.0, 0.0,
-				GridBagConstraints.WEST, GridBagConstraints.NONE, INSETS, 0, 0));
-		inputPanel.add(pathJsonata, new GridBagConstraints(x++, y, 1, 1, 1.0, 1.0,
-				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-		inputPanel.add(changePathJsonata, new GridBagConstraints(x++, y, 1, 1, 0.0, 0.0,
-				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-		inputPanel.add(loadJsonata, new GridBagConstraints(x++, y++, 1, 1, 0.0, 0.0,
-				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-
-//		x = 1;
-//		inputPanel.add(new JLabel("Driver File Path Firefox"), new GridBagConstraints(x++, y, 1, 1, 0.0, 0.0,
-//				GridBagConstraints.WEST, GridBagConstraints.NONE, INSETS, 0, 0));
-//		inputPanel.add(driverFilePathFi, new GridBagConstraints(x++, y, 1, 1, 1.0, 1.0,
-//				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-//		inputPanel.add(changeDriverFilePathFi, new GridBagConstraints(x++, y++, 1, 1, 0.0, 0.0,
-//				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-//
-//		x = 1;
-//		inputPanel.add(new JLabel("Driver File Path Chrome"), new GridBagConstraints(x++, y, 1, 1, 0.0, 0.0,
-//				GridBagConstraints.WEST, GridBagConstraints.NONE, INSETS, 0, 0));
-//		inputPanel.add(driverFilePathCh, new GridBagConstraints(x++, y, 1, 1, 1.0, 1.0,
-//				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-//		inputPanel.add(changeDriverFilePathCh, new GridBagConstraints(x++, y++, 1, 1, 0.0, 0.0,
-//				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-//
-//		x = 1;
-//		inputPanel.add(new JLabel("Download Folder"), new GridBagConstraints(x++, y, 1, 1, 0.0, 0.0,
-//				GridBagConstraints.WEST, GridBagConstraints.NONE, INSETS, 0, 0));
-//		inputPanel.add(downloadFolder, new GridBagConstraints(x++, y, 1, 1, 1.0, 1.0,
-//				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-//		inputPanel.add(changeDownloadFolder, new GridBagConstraints(x++, y++, 1, 1, 0.0, 0.0,
-//				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-//
-//		x = 1;
-//		inputPanel.add(new JLabel("EAR Upload Mode"), new GridBagConstraints(x++, y, 1, 1, 0.0, 0.0,
-//				GridBagConstraints.WEST, GridBagConstraints.NONE, INSETS, 0, 0));
-//		inputPanel.add(changeDownloadFolder, new GridBagConstraints(x++, y++, 1, 1, 0.0, 0.0,
-//				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-//
-//		inputPanel.add(new JLabel("Driver Max Wait Seconds General"), new GridBagConstraints(1, y, 1, 1, 0.0, 0.0,
-//				GridBagConstraints.WEST, GridBagConstraints.NONE, INSETS, 0, 0));
-//		inputPanel.add(driverMaxWaitSeconds, new GridBagConstraints(2, y++, 1, 1, 1.0, 1.0,
-//				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-//
-//		inputPanel.add(new JLabel("Driver Max Wait Seconds Page Load"), new GridBagConstraints(1, y, 1, 1, 0.0, 0.0,
-//				GridBagConstraints.WEST, GridBagConstraints.NONE, INSETS, 0, 0));
-//		inputPanel.add(driverMaxWaitSecondsPageLoad, new GridBagConstraints(2, y++, 1, 1, 1.0, 1.0,
-//				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-//
-//		inputPanel.add(new JLabel("Driver Max Wait Seconds Script"), new GridBagConstraints(1, y, 1, 1, 0.0, 0.0,
-//				GridBagConstraints.WEST, GridBagConstraints.NONE, INSETS, 0, 0));
-//		inputPanel.add(driverMaxWaitSecondsScript, new GridBagConstraints(2, y++, 1, 1, 1.0, 1.0,
-//				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-//
-//		inputPanel.add(new JLabel("Driver Max Wait Seconds Start"), new GridBagConstraints(1, y, 1, 1, 0.0, 0.0,
-//				GridBagConstraints.WEST, GridBagConstraints.NONE, INSETS, 0, 0));
-//		inputPanel.add(driverMaxWaitSecondsStart, new GridBagConstraints(2, y++, 1, 1, 1.0, 1.0,
-//				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-//
-//		inputPanel.add(new JLabel("Driver Max Wait Seconds Stop"), new GridBagConstraints(1, y, 1, 1, 0.0, 0.0,
-//				GridBagConstraints.WEST, GridBagConstraints.NONE, INSETS, 0, 0));
-//		inputPanel.add(driverMaxWaitSecondsStop, new GridBagConstraints(2, y++, 1, 1, 1.0, 1.0,
-//				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-//
-//		inputPanel.add(new JLabel("Driver Max Wait Seconds Upload"), new GridBagConstraints(1, y, 1, 1, 0.0, 0.0,
-//				GridBagConstraints.WEST, GridBagConstraints.NONE, INSETS, 0, 0));
-//		inputPanel.add(driverMaxWaitSecondsUpload, new GridBagConstraints(2, y++, 1, 1, 1.0, 1.0,
-//				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-//
-//		inputPanel.add(new JLabel("Experts Mode"), new GridBagConstraints(1, y, 1, 1, 0.0, 0.0,
-//				GridBagConstraints.WEST, GridBagConstraints.NONE, INSETS, 0, 0));
-//		inputPanel.add(advancedModeCheckBox, new GridBagConstraints(2, y++, 1, 1, 1.0, 1.0,
-//				GridBagConstraints.EAST, GridBagConstraints.HORIZONTAL, INSETS, 0, 0));
-
-		buttonsPanel.add(okButton, new GridBagConstraints(x++, 1, 1, 1, 0.0, 0.0,
-				GridBagConstraints.WEST, GridBagConstraints.NONE, INSETS, 0, 0));
-		buttonsPanel.add(cancelButton, new GridBagConstraints(x++, 1, 1, 1, 0.0, 0.0,
-				GridBagConstraints.WEST, GridBagConstraints.NONE, INSETS, 0, 0));
-
-		southPanel.add(buttonsPanel, BorderLayout.SOUTH);
-
-		frame.add(inputPanel, BorderLayout.NORTH);
-		frame.add(southPanel, BorderLayout.SOUTH);
-	}
-
-	private void uiSetListeners() {
-		uiListenerFrame();
-//		uiListenerDriverKind();
-		uiListenerChangePathInput();
-		uiListenerLoadInput();
-		uiListenerChangePathJsonata();
-		uiListenerLoadJsonata();
-//		uiListenerChangeDriverFilePathFi();
-//		uiListenerChangeDriverFilePathCh();
-		uiListenerOkButton();
-		uiListenerCancelButton();
-	}
-
-	private void uiListenerFrame() {
-		frame.addWindowListener(new WindowAdapter() {
-			public void windowClosing(WindowEvent event) {
-				cancel();
-			}
-		});
-	}
-
-//	private void uiListenerDriverKind() {
-//		driverKind.addActionListener(new ActionListener() {
-//			@Override
-//			public void actionPerformed(ActionEvent e) {
-//				// switchDriverKind(e);
-//			}
-//		});
-//	}
-
-	private void uiListenerChangePathInput() {
-		changePathInput.addActionListener(new ActionListener() {
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				changeFilePath(pathInput, "JSON or XML", ".json", ".xml");
-			}
-		});
-	}
-
-	private void uiListenerLoadInput() {
-		loadInput.addActionListener(new ActionListener() {
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				ui.loadInput(new File(pathInput.getText()).toPath());
-			}
-		});
-	}
-
-	private void uiListenerChangePathJsonata() {
-		changePathJsonata.addActionListener(new ActionListener() {
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				changeFilePath(pathJsonata, "JSONata mapping description", ".jsonata");
-			}
-		});
-	}
-
-	private void uiListenerLoadJsonata() {
-		loadJsonata.addActionListener(new ActionListener() {
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				ui.loadJsoanata(new File(pathJsonata.getText()).toPath());
-			}
-		});
-	}
-
-//	private void uiListenerChangeDriverFilePathFi() {
-//		changeDriverFilePathFi.addActionListener(new ActionListener() {
-//			@Override
-//			public void actionPerformed(ActionEvent e) {
-//				changeDriverFilePath(driverFilePathFi);
-//			}
-//		});
-//	}
-//
-//	private void uiListenerChangeDriverFilePathCh() {
-//		changeDriverFilePathCh.addActionListener(new ActionListener() {
-//			@Override
-//			public void actionPerformed(ActionEvent e) {
-//				changeDriverFilePath(driverFilePathCh);
-//			}
-//		});
-//	}
-
-	private void uiListenerOkButton() {
-		okButton.addActionListener(new ActionListener() {
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				applyAndClose();
-			}
-		});
-	}
-
-	private void uiListenerCancelButton() {
-		cancelButton.addActionListener(new ActionListener() {
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				cancel();
-			}
-		});
-	}
-}
