+++ src/test/resources/exerciser/address.json
@@ -1,39 +1,53 @@
 {
-  "FirstName" : "Fred",
-  "Surname" : "Smith",
-  "Age" : "28",
-  "Address" : {
-    "Street" : "Hursley Park",
-    "City" : "Winchester",
-    "Postcode" : "SO21 2JN"
+  "FirstName": "Fred",
+  "Surname": "Smith",
+  "Age": 28,
+  "Address": {
+    "Street": "Hursley Park",
+    "City": "Winchester",
+    "Postcode": "SO21 2JN"
   },
-  "Phone" : [ {
-    "type" : "home",
-    "number" : "0203 544 1234"
-  }, {
-    "type" : "office",
-    "number" : "01962 001234"
-  }, {
-    "type" : "office",
-    "number" : "01962 001235"
-  }, {
-    "type" : "mobile",
-    "number" : "077 7700 1234"
-  } ],
-  "Email" : [ {
-    "type" : "office",
-    "address" : [ "fred.smith@my-work.com", "fsmith@my-work.com" ]
-  }, {
-    "type" : "home",
-    "address" : [ "freddy@my-social.com", "frederic.smith@very-serious.com" ]
-  } ],
-  "Other" : {
-    "Over_18" : "true",
-    "Misc" : "",
-    "Alternative.Address" : {
-      "Street" : "Brick Lane",
-      "City" : "London",
-      "Postcode" : "E1 6RF"
+  "Phone": [
+    {
+      "type": "home",
+      "number": "0203 544 1234"
+    },
+    {
+      "type": "office",
+      "number": "01962 001234"
+    },
+    {
+      "type": "office",
+      "number": "01962 001235"
+    },
+    {
+      "type": "mobile",
+      "number": "077 7700 1234"
+    }
+  ],
+  "Email": [
+    {
+      "type": "office",
+      "address": [
+        "fred.smith@my-work.com",
+        "fsmith@my-work.com"
+      ]
+    },
+    {
+      "type": "home",
+      "address": [
+        "freddy@my-social.com",
+        "frederic.smith@very-serious.com"
+      ]
+    }
+  ],
+  "Other": {
+    "Over 18 ?": true,
+    "Misc": null,
+    "Alternative.Address": {
+      "Street": "Brick Lane",
+      "City": "London",
+      "Postcode": "E1 6RF"
     }
   }
 }
