+++ src/test/resources/exerciser/invoice.jsonata
@@ -1,0 +1,1 @@
+$sum(Account.Order.Product.(Price * Quantity))
