+++ src/test/resources/exerciser/schema.jsonata
@@ -1,0 +1,1 @@
+**.properties ~> $keys()
